# openapi.json
{
  "components": {
    "schemas": {
...
      "ContadoresSalida": {
        "properties": {
          "adjuntos_descargados": {
            "default": 0,
            "title": "Adjuntos Descargados",
            "type": "integer"
          },
          "adjuntos_duplicados": {
            "default": 0,
            "title": "Adjuntos Duplicados",
            "type": "integer"
          },
          "adjuntos_rechazados": {
            "default": 0,
            "title": "Adjuntos Rechazados",
            "type": "integer"
          },
          "errores": {
            "default": 0,
            "title": "Errores",
            "type": "integer"
          },
          "mensajes_con_adjuntos": {
            "default": 0,
            "title": "Mensajes Con Adjuntos",
            "type": "integer"
          },
          "mensajes_revisados": {
            "default": 0,
            "title": "Mensajes Revisados",
            "type": "integer"
          }
        },
        "title": "ContadoresSalida",
        "type": "object"
      },
...
      "ErrorDeProcesamientoSalida": {
        "properties": {
          "codigo": {
            "title": "Codigo",
            "type": "string"
          },
          "etapa": {
            "title": "Etapa",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mensaje": {
            "title": "Mensaje",
            "type": "string"
          },
          "ocurrido_en": {
            "format": "date-time",
            "title": "Ocurrido En",
            "type": "string"
          },
          "reintentable": {
            "title": "Reintentable",
            "type": "boolean"
          },
          "trabajo_id": {
            "format": "uuid",
            "title": "Trabajo Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trabajo_id",
          "etapa",
          "codigo",
          "mensaje",
          "reintentable",
          "ocurrido_en"
        ],
        "title": "ErrorDeProcesamientoSalida",
        "type": "object"
      },
      "EscaneoSalida": {
        "properties": {
          "codigo_de_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo De Error"
          },
          "contadores": {
            "$ref": "#/components/schemas/ContadoresSalida"
          },
          "encolado_en": {
            "format": "date-time",
            "title": "Encolado En",
            "type": "string"
          },
          "estado": {
            "title": "Estado",
            "type": "string"
          },
          "fase": {
            "title": "Fase",
            "type": "string"
          },
          "finalizado_en": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalizado En"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "iniciado_en": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iniciado En"
          },
          "mensaje_de_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mensaje De Error"
          },
          "progreso_porcentaje": {
            "title": "Progreso Porcentaje",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "estado",
          "fase",
          "progreso_porcentaje",
          "contadores",
          "codigo_de_error",
          "mensaje_de_error",
          "encolado_en",
          "iniciado_en",
          "finalizado_en"
        ],
        "title": "EscaneoSalida",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IniciarEscaneoEntrada": {
        "additionalProperties": false,
        "properties": {
          "carpeta": {
            "default": "INBOX",
            "maxLength": 120,
            "pattern": "^[\\w\\-/ ]+$",
            "title": "Carpeta",
            "type": "string"
          },
          "conexion_id": {
            "format": "uuid",
            "title": "Conexion Id",
            "type": "string"
          },
          "desde": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desde"
          },
          "hasta": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hasta"
          },
          "limite_de_mensajes": {
            "default": 100,
            "maximum": 50000.0,
            "minimum": 1.0,
            "title": "Limite De Mensajes",
            "type": "integer"
          }
        },
        "required": [
          "conexion_id"
        ],
        "title": "IniciarEscaneoEntrada",
        "type": "object"
      },
...
      "Respuesta_EscaneoSalida_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EscaneoSalida"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Respuesta[EscaneoSalida]",
        "type": "object"
      },
...
      "Respuesta_list_ErrorDeProcesamientoSalida__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ErrorDeProcesamientoSalida"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Respuesta[list[ErrorDeProcesamientoSalida]]",
        "type": "object"
      },
      "Respuesta_list_EscaneoSalida__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EscaneoSalida"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Respuesta[list[EscaneoSalida]]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Ingesta de adjuntos de correo y extraccion de datos tributarios. Todas las rutas requieren un token OIDC salvo las sondas de salud.",
    "title": "Automatizacion de Correos API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scans": {
      "get": {
        "operationId": "listar_api_v1_scans_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_list_EscaneoSalida__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar",
        "tags": [
          "escaneos"
        ]
      },
      "post": {
        "description": "Encola un escaneo. Idempotente si se envia `Idempotency-Key`.",
        "operationId": "iniciar_api_v1_scans_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IniciarEscaneoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_EscaneoSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Iniciar",
        "tags": [
          "escaneos"
        ]
      }
    },
    "/api/v1/scans/{trabajo_id}": {
      "get": {
        "operationId": "obtener_api_v1_scans__trabajo_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trabajo_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trabajo Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_EscaneoSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Obtener",
        "tags": [
          "escaneos"
        ]
      }
    },
    "/api/v1/scans/{trabajo_id}/cancel": {
      "post": {
        "operationId": "cancelar_api_v1_scans__trabajo_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "trabajo_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trabajo Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_EscaneoSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancelar",
        "tags": [
          "escaneos"
        ]
      }
    },
    "/api/v1/scans/{trabajo_id}/errors": {
      "get": {
        "operationId": "errores_del_escaneo_api_v1_scans__trabajo_id__errors_get",
        "parameters": [
          {
            "in": "path",
            "name": "trabajo_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trabajo Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_list_ErrorDeProcesamientoSalida__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Errores Del Escaneo",
        "tags": [
          "escaneos"
        ]
      }
    },
    "/api/v1/scans/{trabajo_id}/stream": {
      "get": {
        "description": "Progreso en vivo por Server-Sent Events.\n\nLa autorizacion se comprueba ANTES de abrir el stream: una vez\nestablecido, la conexion queda abierta y ya no hay donde devolver un\n403 que el cliente interprete correctamente.",
        "operationId": "transmitir_progreso_api_v1_scans__trabajo_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "trabajo_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trabajo Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transmitir Progreso",
        "tags": [
          "escaneos"
        ]
      }
    },
...
  }
}